+++ client/src/components/Eduprocess.vue
@@ -1,19 +1,20 @@
 <template>
-  <div>
-    <div>
-      <h3>Навчально-методична робота</h3>
-      <label for="amountnm">Назва. Кількісна характеристика</label>
-      <input v-model="amountnm" id="amountnm" type="text">
-      <label for="termnm">Термін виконання</label>
-      <input v-model="termnm" id="termnm" type="text">
-      <button @click.prevent="create('nm', amountnm, termnm)">Додати</button>
+  <div class="container" style="max-width: 80%">
+    <div class="mb-5">
+      <h3 class="mb-4">Навчально-методична робота</h3>
+      <div class="mb-4 form-group d-flex justify-content-between align-items-center">
+        <label  class="my-auto" for="amountnm">Назва. Кількісна характеристика:</label>
+        <input class="form-control" v-model="amountnm" id="amountnm" type="text">
+        <label  class="my-auto" for="termnm">Термін виконання:</label>
+        <input class="form-control" v-model="termnm" id="termnm" type="text">
+        <button class="btn btn-primary" @click.prevent="create('nm', amountnm, termnm)">Додати</button>
+      </div>
       <table class="table table-bordered">
         <tr>
           <th>Назва. Кількісна характеристика</th>
           <th>Термін виконання</th>
         </tr>
         <tr
-          @dblclick="deletes(item._id)"
           v-for="(item, index) in list"
           :item="item"
           :index="index"
@@ -22,24 +23,26 @@
           <template v-if="item.type=='nm'">
             <td>{{item.amount}}</td>
             <td>{{item.term}}</td>
+            <td style="cursor: pointer; text-align: center" @click.prevent="deletes(item._id)"><img   src="img/delete.svg" width="15px"  alt=""></td>
           </template>
         </tr>
       </table>
     </div>
-    <div>
-      <h3>Науково-дослідна робота</h3>
-      <label for="amountnd">Назва. Кількісна характеристика</label>
-      <input v-model="amountnd" id="amountnd" type="text">
-      <label for="termnd">Термін виконання</label>
-      <input v-model="termnd" id="termnd" type="text">
-      <button @click.prevent="create('nd', amountnd, termnd)">Додати</button>
+    <div class="mb-5">
+      <h3 class="mb-4">Науково-дослідна робота</h3>
+      <div class="mb-4 form-group d-flex justify-content-between align-items-center">
+        <label class="my-auto" for="amountnd">Назва. Кількісна характеристика</label>
+        <input class="form-control" v-model="amountnd" id="amountnd" type="text">
+        <label class="my-auto" for="termnd">Термін виконання</label>
+        <input class="form-control" v-model="termnd" id="termnd" type="text">
+        <button class="btn btn-primary" @click.prevent="create('nd', amountnd, termnd)">Додати</button>
+      </div>
       <table class="table table-bordered">
         <tr>
           <th>Назва. Кількісна характеристика</th>
           <th>Термін виконання</th>
         </tr>
         <tr
-          @dblclick="deletes(item._id)"
           v-for="(item, index) in list"
           :item="item"
           :index="index"
@@ -48,24 +51,27 @@
           <template v-if="item.type=='nd'">
             <td>{{item.amount}}</td>
             <td>{{item.term}}</td>
+            <td style="cursor: pointer; text-align: center" @click.prevent="deletes(item._id)"><img   src="img/delete.svg" width="15px"  alt=""></td>
           </template>
         </tr>
       </table>
     </div>
-    <div>
-      <h3>Організаційно-методична робота</h3>
-      <label for="amountom">Назва. Кількісна характеристика</label>
-      <input v-model="amountom" id="amountom" type="text">
-      <label for="termom">Термін виконання</label>
-      <input v-model="termom" id="termom" type="text">
-      <button @click.prevent="create('om', amountom, termom)">Додати</button>
+    <div class="mb-5">
+      <h3 class="mb-4">Організаційно-методична робота</h3>
+      <div class="mb-4 form-group d-flex justify-content-between align-items-center">
+        <label class="my-auto" for="amountom">Назва. Кількісна характеристика</label>
+        <input class="form-control" v-model="amountom" id="amountom" type="text">
+        <label class="my-auto" for="termom">Термін виконання</label>
+        <input class="form-control" v-model="termom" id="termom" type="text">
+        <button class="btn btn-primary" @click.prevent="create('om', amountom, termom)">Додати</button>
+      </div>
       <table class="table table-bordered">
         <tr>
           <th>Назва. Кількісна характеристика</th>
           <th>Термін виконання</th>
         </tr>
         <tr
-          @dblclick="deletes(item._id)"
+
           v-for="(item, index) in list"
           :item="item"
           :index="index"
@@ -74,24 +80,26 @@
           <template v-if="item.type=='om'">
             <td>{{item.amount}}</td>
             <td>{{item.term}}</td>
+            <td style="cursor: pointer; text-align: center" @click.prevent="deletes(item._id)"><img   src="img/delete.svg" width="15px"  alt=""></td>
           </template>
         </tr>
       </table>
     </div>
-    <div>
-      <h3>Громадська робота та iншi види робiт</h3>
-      <label for="amounthr">Назва. Кількісна характеристика</label>
-      <input v-model="amounthr" id="amounthr" type="text">
-      <label for="termhr">Термін виконання</label>
-      <input v-model="termhr" id="termhr" type="text">
-      <button @click.prevent="create('hr', amounthr, termhr)">Додати</button>
+    <div class="mb-5">
+      <h3 class="mb-4">Громадська робота та iншi види робiт</h3>
+      <div class="mb-4 form-group d-flex justify-content-between align-items-center">
+        <label class="my-auto" for="amounthr">Назва. Кількісна характеристика</label>
+        <input class="form-control" v-model="amounthr" id="amounthr" type="text">
+        <label class="my-auto" for="termhr">Термін виконання</label>
+        <input class="form-control" v-model="termhr" id="termhr" type="text">
+        <button class="btn btn-primary" @click.prevent="create('hr', amounthr, termhr)">Додати</button>
+      </div>
       <table class="table table-bordered">
         <tr>
           <th>Назва. Кількісна характеристика</th>
           <th>Термін виконання</th>
         </tr>
         <tr
-          @dblclick="deletes(item._id)"
           v-for="(item, index) in list"
           :item="item"
           :index="index"
@@ -100,24 +108,26 @@
           <template v-if="item.type=='hr'">
             <td>{{item.amount}}</td>
             <td>{{item.term}}</td>
+            <td style="cursor: pointer; text-align: center" @click.prevent="deletes(item._id)"><img   src="img/delete.svg" width="15px"  alt=""></td>
           </template>
         </tr>
       </table>
     </div>
-    <div>
-      <h3>Виховна робота</h3>
-      <label for="amounted">Назва. Кількісна характеристика</label>
-      <input v-model="amounted" id="amounted" type="text">
-      <label for="termed">Термін виконання</label>
-      <input v-model="termed" id="termed" type="text">
-      <button @click.prevent="create('ed', amounted, termed)">Додати</button>
+    <div class="mb-5">
+      <h3 class="mb-4">Виховна робота</h3>
+      <div class="mb-4 form-group d-flex justify-content-between align-items-center">
+        <label class="my-auto" for="amounted">Назва. Кількісна характеристика</label>
+        <input class="form-control" v-model="amounted" id="amounted" type="text">
+        <label  class="my-auto" for="termed">Термін виконання</label>
+        <input class="form-control" v-model="termed" id="termed" type="text">
+        <button class="btn btn-primary" @click.prevent="create('ed', amounted, termed)">Додати</button>
+      </div>
       <table class="table table-bordered">
         <tr>
           <th>Назва. Кількісна характеристика</th>
           <th>Термін виконання</th>
         </tr>
         <tr
-          @dblclick="deletes(item._id)"
           v-for="(item, index) in list"
           :item="item"
           :index="index"
@@ -126,6 +136,7 @@
           <template v-if="item.type=='ed'">
             <td>{{item.amount}}</td>
             <td>{{item.term}}</td>
+            <td style="cursor: pointer; text-align: center" @click.prevent="deletes(item._id)"><img   src="img/delete.svg" width="15px"  alt=""></td>
           </template>
         </tr>
       </table>
@@ -165,9 +176,17 @@ export default {
       this.list = await eduprocess.getProcess(this.$root.user);
     },
     async deletes(id) {
-      await eduprocess.deleteProcess(id, this.$root.user);
-      this.list = await eduprocess.getProcess(this.$root.user);
+      if(confirm('Ви впевненнi?')){
+        await eduprocess.deleteProcess(id, this.$root.user);
+        this.list = await eduprocess.getProcess(this.$root.user);
+      }
+      
     }
   }
 };
 </script>
+<style scoped>
+    .form-control{
+      width: 30%
+    }
+  </style>
