+++ client/src/components/Methodicals.vue
@@ -1,13 +1,13 @@
 <template>
-  <div>
-    <div>
-    <label for="name">Ф.І.О. викладача:</label><input v-model="name" id="name" type="text">
-    <label for="course">Курс:</label><input v-model="course" id="course" type="number">
-    <label for="semester">Семестр:</label><input v-model="semester" id="semester" type="number">
-    <label for="discipline">Назва дисципліни:</label><input v-model="discipline" id="discipline" type="text">
-    <button @click.prevent="createMeth">Додати</button>
+  <div style="max-width: 80%" class="container w-100">
+    <div v-if="$root.user=='admin'" class="mb-5 form-group d-flex justify-content-between align-items-center">
+      <label class="my-auto" for="name">Ф.І.О. викладача:</label><input class="form-control" v-model="name" id="name" type="text">
+      <label class="my-auto" for="course">Курс:</label><input class="form-control" v-model="course" id="course" type="number">
+      <label class="my-auto" for="semester">Семестр:</label><input class="form-control" v-model="semester" id="semester" type="number">
+      <label class="my-auto" for="discipline">Назва дисципліни:</label><input class="form-control" v-model="discipline" id="discipline" type="text">
+      <button class="btn btn-primary" @click.prevent="createMeth">Додати</button>
     </div>
-    <table class="table">
+    <table class="table table-bordered">
       <tr>
         <th>Ф.І.О. викладача</th>
         <th>Курс</th>
@@ -15,7 +15,7 @@
         <th>Назва дисципліни</th>
       </tr>
     <tr
-        @dblclick="deleteMeth(meth._id)"
+        
         v-for="(meth, index) in list"
         :item="meth"
         :index="index"
@@ -25,6 +25,7 @@
         <td>{{meth.course}}</td>
         <td>{{meth.semester}}</td>
         <td>{{meth.discipline}}</td>
+        <td v-if="$root.user=='admin'" style="cursor: pointer; text-align: center" @click.prevent="deleteMeth(meth._id)"><img   src="img/delete.svg" width="15px"  alt=""></td>
       </tr>
       </table>
   </div>
@@ -57,9 +58,17 @@ export default {
       this.list = await methodicals.getMeth();
     },
     async deleteMeth(id) {
-      await methodicals.deleteMeth(id);
-      this.list = await methodicals.getMeth();
+      if(confirm('Ви впевненнi?')){
+        await methodicals.deleteMeth(id);
+        this.list = await methodicals.getMeth();
+      }
+      
     }
   }
 }
 </script>
+<style scoped>
+  .form-control{
+    width: 15%
+  }
+</style>
