+++ client/src/components/Teachers.vue
@@ -1,13 +1,13 @@
 <template>
-  <div>
+  <div class="container" style="max-width: 80%">
     <h1>Список викладачiв</h1>
-    <div>
-    <label for="name">Ф.І.О. викладача:</label><input v-model="name" id="name" type="text">
-    <label for="login">Логiн:</label><input v-model="login" id="login" type="text">
-    <label for="password">Пароль:</label><input v-model="password" id="password" type="text">
-    <button @click.prevent="createTeach">Додати</button>
+      <div class="mb-5 form-group d-flex justify-content-between align-items-center">
+      <label class="my-auto" for="name">Ф.І.О. викладача:</label><input class="form-control" v-model="name" id="name" type="text">
+      <label class="my-auto" for="login">Логiн:</label><input class="form-control" v-model="login" id="login" type="text">
+      <label class="my-auto" for="password">Пароль:</label><input class="form-control" v-model="password" id="password" type="text">
+      <button class="btn btn-primary" @click.prevent="createTeach">Додати</button>
     </div>
-    <table class="table">
+    <table class="table table-bordered">
       <tr>
         <th>Ф.І.О. викладача</th>
         <th>Логiн</th>
@@ -23,6 +23,7 @@
         <td>{{teach.name}}</td>
         <td>{{teach.login}}</td>
         <td>{{teach.password}}</td>
+        <td v-if="$root.user=='admin'" style="cursor: pointer; text-align: center" @click.prevent="deleteTeach(teach._id, teach.name)"><img   src="img/delete.svg" width="15px"  alt=""></td>
       </tr>
       </table>
   </div>
@@ -50,9 +51,17 @@ export default {
       this.$root.users = await teachers.getTeach();
     },
     async deleteTeach(id, dir) {
-      await teachers.deleteTeach(id, dir);
-      this.$root.users = await teachers.getTeach();
+      if(confirm('Ви впевненнi?')){
+        await teachers.deleteTeach(id, dir);
+        this.$root.users = await teachers.getTeach();
+      }
+      
     }
   }
 }
 </script>
+<style scoped>
+    .form-control{
+      width: 25%
+    }
+  </style>
